requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
sqlalchemy==2.0.23
pyodbc==5.0.1
gunicorn==21.2.0
pydantic==2.5.0
